[MISC] Move configuration sub-sub-module to be a sub-module of search
