+++ include/seqan3/search/configuration/detail.hpp
