+++ include/seqan3/search/configuration/max_error.hpp
@@ -6,7 +6,7 @@
 // -----------------------------------------------------------------------------------------------------
 
 /*!\file
- * \brief Provides the configuration for maximum number of errors in percent of the query length across all error types.
+ * \brief Provides the configuration for maximum number of errors for all error types.
  * \author Christopher Pockrandt <christopher.pockrandt AT fu-berlin.de>
  * \author Rene Rahn <rene.rahn AT fu-berlin.de>
  */
@@ -19,16 +19,14 @@
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
 #include <seqan3/core/algorithm/parameter_pack.hpp>
 #include <seqan3/range/view/slice.hpp>
-#include <seqan3/search/algorithm/configuration/detail.hpp>
-#include <seqan3/search/algorithm/configuration/max_error_common.hpp>
+#include <seqan3/search/configuration/detail.hpp>
+#include <seqan3/search/configuration/max_error_common.hpp>
 #include <seqan3/std/algorithm>
 
 namespace seqan3::search_cfg
 {
-
-/*!\brief A configuration element for the maximum number of errors in percent of the query length across all error types
- *        (mismatches, insertions, deletions). This is an upper bound of errors independent from error rates of
- *        specific error types.
+/*!\brief A configuration element for the maximum number of errors across all error types (mismatches, insertions,
+ *        deletions). This is an upper bound of errors independent from error numbers of specific error types.
  * \details An insertion corresponds to a base inserted into the query that does not occur in the text at the position,
  *          a deletion corresponds to a base deleted from the query sequence that does occur in the indexed text.
  *          Deletions at the beginning and at the end of the sequence are not considered during a search.
@@ -42,13 +40,12 @@ template <typename ...errors_t>
               detail::is_type_specialisation_of_v<std::remove_reference_t<errors_t>, deletion>  ||
               detail::is_type_specialisation_of_v<std::remove_reference_t<errors_t>, insertion>) && ...)
 //!\endcond
-class max_error_rate : public pipeable_config_element<max_error_rate<errors_t...>, std::array<double, 4>>
+class max_error : public pipeable_config_element<max_error<errors_t...>, std::array<uint8_t, 4>>
 {
-
     //!\brief An alias type for the base class.
-    using base_t = pipeable_config_element<max_error_rate<errors_t...>, std::array<double, 4>>;
+    using base_t = pipeable_config_element<max_error<errors_t...>, std::array<uint8_t, 4>>;
 
-    //!\brief Helper function to check valid error rate configuration.
+    //!\brief Helper function to check valid max error configuration.
     template <typename ..._errors_t>
     static constexpr bool check_consistency(_errors_t ...errors)
     {
@@ -76,19 +73,19 @@ public:
 
     //!\privatesection
     //!\brief Internal id to check for consistent configuration settings.
-    static constexpr detail::search_config_id id{detail::search_config_id::max_error_rate};
+    static constexpr detail::search_config_id id{detail::search_config_id::max_error};
 
     //!\publicsection
     /*!\name Constructor, destructor and assignment
      * \brief Defaulted all standard constructor.
      * \{
      */
-    constexpr max_error_rate()                                   noexcept = default; //!< Default constructor.
-    constexpr max_error_rate(max_error_rate const &)             noexcept = default; //!< Copy constructor.
-    constexpr max_error_rate(max_error_rate &&)                  noexcept = default; //!< Move constructor.
-    constexpr max_error_rate & operator=(max_error_rate const &) noexcept = default; //!< Copy assignment.
-    constexpr max_error_rate & operator=(max_error_rate &&)      noexcept = default; //!< Move assignment.
-    ~max_error_rate()                                            noexcept = default; //!< Destructor.
+    constexpr max_error()                              noexcept = default; //!< Default constructor.
+    constexpr max_error(max_error const &)             noexcept = default; //!< Copy constructor.
+    constexpr max_error(max_error &&)                  noexcept = default; //!< Move constructor.
+    constexpr max_error & operator=(max_error const &) noexcept = default; //!< Copy assignment.
+    constexpr max_error & operator=(max_error &&)      noexcept = default; //!< Move assignment.
+    ~max_error()                                       noexcept = default; //!< Destructor.
 
     /*!\brief Constructs the object from a set of error specifiers.
      * \tparam    errors_t A template parameter pack with the error types.
@@ -96,10 +93,8 @@ public:
      *
      * \details
      *
-     * \details
-     *
-     * This configuration can be used to specify the total rates of error types.
-     * It restricts the rates of substitutions, insertions, deletions and total errors within the search to the given
+     * This configuration can be used to specify the total number of error types.
+     * It restricts the number of substitutions, insertions, deletions and total errors within the search to the given
      * values and will behave as follows:
      * |                                  |                              |                           |                        |                       |                                       |                                      |                                   |
      * |----------------------------------|:----------------------------:|:-------------------------:|:----------------------:|:---------------------:|:-------------------------------------:|:------------------------------------:|:---------------------------------:|
@@ -112,9 +107,9 @@ public:
      *
      * ### Example
      *
-     * \include test/snippet/search/configuration_error_rate.cpp
+     * \include test/snippet/search/configuration_error.cpp
      */
-    constexpr max_error_rate(errors_t && ...errors)
+    constexpr max_error(errors_t && ...errors) noexcept
     //!\cond
         requires sizeof...(errors_t) > 0
     //!\endcond
@@ -125,13 +120,6 @@ public:
             base_t::value[remove_cvref_t<decltype(e)>::_id()] = e.get();
         }, std::forward<errors_t>(errors)...);
 
-        // check correct values.
-        std::ranges::for_each(base_t::value, [](auto error_elem)
-        {
-            if (0.0 > error_elem  || error_elem > 1.0)
-                throw std::invalid_argument("Error rates must be between 0 and 1.");
-        });
-
         // Only total is set so we set all other errors to the total limit.
         if constexpr (((std::remove_reference_t<errors_t>::_id() == 0) || ...) && sizeof...(errors) == 1)
         {
@@ -139,23 +127,24 @@ public:
         } // otherwise if total is not set but any other field is set than use total as the sum of all set errors.
         else if constexpr (!((std::remove_reference_t<errors_t>::_id() == 0) || ...) && sizeof...(errors) > 0)
         {
-            base_t::value[0] = std::min(1., ranges::accumulate(base_t::value | view::slice(1, 4), .0));
+            base_t::value[0] = std::min(static_cast<uint8_t>(255), ranges::accumulate(base_t::value | view::slice(1, 4),
+                                                                              static_cast<uint8_t>(0)));
         }
     }
     //!}
 };
 
 /*!\name Type deduction guides
- * \relates seqan3::search_cfg::max_error_rate
+ * \relates seqan3::search_cfg::max_error
  * \{
  */
 
 //!\brief Deduces empty list of error specifiers.
-max_error_rate() -> max_error_rate<>;
+max_error() -> max_error<>;
 
 //!\brief Deduces template arguments from the passed error specifiers.
 template <typename ...errors_t>
-max_error_rate(errors_t && ...) -> max_error_rate<remove_cvref_t<errors_t>...>;
+max_error(errors_t && ...) -> max_error<remove_cvref_t<errors_t>...>;
 //!\}
 
 } // namespace seqan3::search_cfg
