+++ include/seqan3/search/configuration/max_error_common.hpp
