+++ include/seqan3/search/configuration/mode.hpp
@@ -17,7 +17,7 @@
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
 #include <seqan3/core/type_traits/basic.hpp>
-#include <seqan3/search/algorithm/configuration/detail.hpp>
+#include <seqan3/search/configuration/detail.hpp>
 
 /*!\addtogroup search
  * \{
