+++ test/snippet/search/configuration_default.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/search/algorithm/configuration/all.hpp>
+#include <seqan3/search/configuration/all.hpp>
 
 int main()
 {
